EventHub: add EventHubPixelFiring protocol and move FiredPixel to its own file
